scenario6/requirements.txt
# Scenario 6 — PersonaPlex Real-Time Voice Conversation
# Install PyTorch with CUDA first:
#   pip install torch --index-url https://download.pytorch.org/whl/cu121
# Then install the vendored moshi package:
#   pip install -e ./third_party/moshi
# Then install these remaining dependencies:

-e ./third_party/moshi

fastapi
uvicorn[standard]
python-dotenv
aiohttp>=3.10.5,<3.11
huggingface-hub>=0.24
websockets
